Updated ComponentManager<T> to store T*s
ComponentManager<T> would previously store Ts, meaning that componets
would actually be moved in memory when the manager was rearranging its
list of active components. Both the potential expense of this and the
inability to guarantee the validity of component pointers in between
tick calls made this solution too unattractive. Storing T*s means that
the componets themselves don't move in memory.