Major EntityManager restructure; systems to follow
Moved to a system where entities are represented by a 64-bit number,
where the MSB 32 bits are a unique ID, the next 16 bits are zeroed out
(a non-zero value here indicates a component type), and a final 16 bits
for use of a "version number", so that after deleting and recreating an
entity of the same ID, there will be a way of disambiguating between
handles to the previous entity and those of the new one.

Architectural changes were made at the EntityManager level to deal with
entities of this new form, and ComponentManagers were changed likewise.
With this new architecture, the ComponentManagers no longer need a hash
table (yay!).

Though not currently supporting concurrency, the system could be fairly
easily modified to use locks in order to allow for thread-safe access.

Signed-off-by: Khouderchah-Alex <alex.khouderchah@gmail.com>