Sound Implementation. Switched to Component System
Buffer and Sound classes are gone. They are replaced with SoundSystem
and SoundComponent classes. SoundSystem will be only responsible for
setting up the context and the device. I spent nearly 5 hours and could
not find a reliable way for SoundSystem to handle playing/pausing the
sound. Best solution was to handle pla/pause by SoundContext.

NOTE: Sound poisition functionality of SoundSystem is not yet
implemented.

For now, we have a working background sound and sound is played when you
press the jump button. This sound is activated in Action_Jump class as
it suppose to be.