Fixed the sound system camera position implementation and added collision(Collision code is just a prototype and is not the final version)
In order to be able to get the current camera position and "AT" and "UP"
components in SoundSystem, I created a static pointer to current camera
in CameraComponent. You can set the current camera by calling the
SetCurrentCamera method, and you can get the current camera by calling
the static method GetCurrnetCamera.

For the collision, you can toggle the hitbox view by pressing the "O"
Button on keyboard.

IMPORTANT: Collision system is nowhere near to being complete, this is
just a mockup code to get me starting. Also, the program currently
contains a bug, where upon exiting the program, we get a memory access
violation due to the code that handles the event that toggles collision
box view. I am currently working on fixing this bug.