Full Physics engine overhaul
I am completely changing the physicsComponent in order to be able to use
Bullet Physics API. So far I successfully integrated the API into our
project and was able to make the collisions work. Program is still buggy
since non-static colliding objects can bounce around from the impact.
Also, this change makes the way I was drawing the hit boxes impossible,
so I need to come up with a new way to make hit boxes visible. This will
be very easy after the code for loading and using 3D shapes is ready.