The new physics system should be ready for push.
I fixed all the bugs and implemented easy ways to manipulate physical
objects trough CollisionComponent. Note that CollisionComponent might
not be a suitable name for that  module since it actually adds physical
bodies rather than collision for the bodies. Physics component now
relies on the above mentioned bodies in order to be able to do anything.